refactors provider configs into separate package to avoid circular dependency b/w metadata and provider packages
