+++ provider/cassandra.go
@@ -9,6 +9,7 @@ import (
 	"encoding/json"
 	"fmt"
 
+	pc "github.com/featureform/provider/provider_config"
 	"github.com/gocql/gocql"
 	sn "github.com/mrz1836/go-sanitize"
 )
@@ -37,8 +38,8 @@ type cassandraOnlineTable struct {
 	valueType ValueType
 }
 
-func cassandraOnlineStoreFactory(serialized SerializedConfig) (Provider, error) {
-	cassandraConfig := &CassandraConfig{}
+func cassandraOnlineStoreFactory(serialized pc.SerializedConfig) (Provider, error) {
+	cassandraConfig := &pc.CassandraConfig{}
 	if err := cassandraConfig.Deserialize(serialized); err != nil {
 		return nil, err
 	}
@@ -49,7 +50,7 @@ func cassandraOnlineStoreFactory(serialized SerializedConfig) (Provider, error) 
 	return NewCassandraOnlineStore(cassandraConfig)
 }
 
-func NewCassandraOnlineStore(options *CassandraConfig) (*cassandraOnlineStore, error) {
+func NewCassandraOnlineStore(options *pc.CassandraConfig) (*cassandraOnlineStore, error) {
 	cassandraCluster := gocql.NewCluster(options.Addr)
 	cassandraCluster.Authenticator = gocql.PasswordAuthenticator{
 		Username: options.Username,
