+++ provider/postgres.go
@@ -2,11 +2,11 @@ package provider
 
 import (
 	"database/sql"
-	"encoding/json"
 	"fmt"
 	"strings"
 	"time"
 
+	pc "github.com/featureform/provider/provider_config"
 	_ "github.com/lib/pq"
 )
 
@@ -21,32 +21,8 @@ const (
 	pgTimestamp                    = "timestamp with time zone"
 )
 
-type PostgresConfig struct {
-	Host     string `json:"Host"`
-	Port     string `json:"Port"`
-	Username string `json:"Username"`
-	Password string `json:"Password"`
-	Database string `json:"Database"`
-}
-
-func (pg *PostgresConfig) Deserialize(config SerializedConfig) error {
-	err := json.Unmarshal(config, pg)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (pg *PostgresConfig) Serialize() []byte {
-	conf, err := json.Marshal(pg)
-	if err != nil {
-		panic(err)
-	}
-	return conf
-}
-
-func postgresOfflineStoreFactory(config SerializedConfig) (Provider, error) {
-	sc := PostgresConfig{}
+func postgresOfflineStoreFactory(config pc.SerializedConfig) (Provider, error) {
+	sc := pc.PostgresConfig{}
 	if err := sc.Deserialize(config); err != nil {
 		return nil, fmt.Errorf("invalid postgres config: %v", config)
 	}
