+++ provider/provider_config/postgres_config.go
@@ -1,0 +1,27 @@
+package provider_config
+
+import "encoding/json"
+
+type PostgresConfig struct {
+	Host     string `json:"Host"`
+	Port     string `json:"Port"`
+	Username string `json:"Username"`
+	Password string `json:"Password"`
+	Database string `json:"Database"`
+}
+
+func (pg *PostgresConfig) Deserialize(config SerializedConfig) error {
+	err := json.Unmarshal(config, pg)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func (pg *PostgresConfig) Serialize() []byte {
+	conf, err := json.Marshal(pg)
+	if err != nil {
+		panic(err)
+	}
+	return conf
+}
