+++ provider/provider_config/provider_config.go
@@ -1,0 +1,3 @@
+package provider_config
+
+type SerializedConfig []byte
