+++ provider/redis.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/go-redis/redis/v8"
 	"time"
+
+	pc "github.com/featureform/provider/provider_config"
+	"github.com/go-redis/redis/v8"
 )
 
 type redisTableKey struct {
@@ -23,8 +25,8 @@ type redisOnlineStore struct {
 	BaseProvider
 }
 
-func redisOnlineStoreFactory(serialized SerializedConfig) (Provider, error) {
-	redisConfig := &RedisConfig{}
+func redisOnlineStoreFactory(serialized pc.SerializedConfig) (Provider, error) {
+	redisConfig := &pc.RedisConfig{}
 	if err := redisConfig.Deserialize(serialized); err != nil {
 		return nil, err
 	}
@@ -34,7 +36,7 @@ func redisOnlineStoreFactory(serialized SerializedConfig) (Provider, error) {
 	return NewRedisOnlineStore(redisConfig), nil
 }
 
-func NewRedisOnlineStore(options *RedisConfig) *redisOnlineStore {
+func NewRedisOnlineStore(options *pc.RedisConfig) *redisOnlineStore {
 	redisOptions := &redis.Options{
 		Addr:     options.Addr,
 		Password: options.Password,
