+++ provider/sql.go
@@ -13,6 +13,7 @@ import (
 	"strings"
 	"time"
 
+	pc "github.com/featureform/provider/provider_config"
 	db "github.com/jackc/pgx/v4"
 	sf "github.com/snowflakedb/gosnowflake"
 )
@@ -22,7 +23,7 @@ func sanitize(ident string) string {
 }
 
 type SQLOfflineStoreConfig struct {
-	Config        SerializedConfig
+	Config        pc.SerializedConfig
 	ConnectionURL string
 	Driver        string
 	ProviderType  Type
